Move CheckExecutable into ee/tuf
