+++ ee/tuf/util_darwin.go
@@ -4,6 +4,8 @@
 package tuf
 
 import (
+	"errors"
+	"fmt"
 	"os"
 	"path/filepath"
 )
@@ -26,4 +28,26 @@ func executableLocation(updateDirectory string, binary autoupdatableBinary) stri
 	default:
 		return ""
 	}
+}
+
+// checkExecutablePermissions checks whether a specific file looks
+// like it's executable. This is used in evaluating whether something
+// is an updated version.
+func checkExecutablePermissions(potentialBinary string) error {
+	if potentialBinary == "" {
+		return errors.New("empty string isn't executable")
+	}
+	stat, err := os.Stat(potentialBinary)
+	switch {
+	case os.IsNotExist(err):
+		return errors.New("no such file")
+	case err != nil:
+		return fmt.Errorf("statting file: %w", err)
+	case stat.IsDir():
+		return errors.New("is a directory")
+	case stat.Mode()&0111 == 0:
+		return errors.New("not executable")
+	}
+
+	return nil
 }
