+++ pkg/autoupdate/findnew_test.go
@@ -365,74 +365,6 @@ func copyFile(dstPath, srcPath string, truncate bool) error {
 	return nil
 }
 
-func TestCheckExecutable(t *testing.T) {
-	t.Parallel()
-
-	// We need to run this from a temp dir, else the early return
-	// from matching os.Executable bypasses the point of this.
-	tmpDir, binaryName := setupTestDir(t, executableUpdates)
-	targetExe := filepath.Join(tmpDir, binaryName)
-
-	var tests = []struct {
-		testName    string
-		expectedErr bool
-	}{
-		{
-			testName:    "exit0",
-			expectedErr: false,
-		},
-		{
-			testName:    "exit1",
-			expectedErr: false,
-		},
-		{
-			testName:    "exit2",
-			expectedErr: false,
-		},
-		{
-			testName:    "sleep",
-			expectedErr: true,
-		},
-	}
-
-	for _, tt := range tests { // nolint:paralleltest
-		tt := tt
-		t.Run(tt.testName, func(t *testing.T) {
-			err := CheckExecutable(context.TODO(), targetExe, "-test.run=TestHelperProcess", "--", tt.testName)
-			if tt.expectedErr {
-				require.Error(t, err, tt.testName)
-
-				// As a bonus, we can test that if we call os.Args[0],
-				// we still don't get an error. This is because we
-				// trigger the match against os.Executable and don't
-				// invoked. This is here, and not a dedicated test,
-				// because we ensure the same test arguments.
-				require.NoError(t, CheckExecutable(context.TODO(), os.Args[0], "-test.run=TestHelperProcess", "--", tt.testName), "calling self with %s", tt.testName)
-			} else {
-				require.NoError(t, err, tt.testName)
-			}
-		})
-
-	}
-}
-
-func TestCheckExecutableTruncated(t *testing.T) {
-	t.Parallel()
-
-	// First make a broken truncated binary. Lots of setup for this.
-	truncatedBinary, err := os.CreateTemp("", "test-autoupdate-check-executable-truncation")
-	require.NoError(t, err, "make temp file")
-	defer os.Remove(truncatedBinary.Name())
-	truncatedBinary.Close()
-
-	require.NoError(t, copyFile(truncatedBinary.Name(), os.Args[0], true), "copy executable")
-	require.NoError(t, os.Chmod(truncatedBinary.Name(), 0755))
-
-	require.Error(t,
-		CheckExecutable(context.TODO(), truncatedBinary.Name(), "-test.run=TestHelperProcess", "--", "exit0"),
-		"truncated binary")
-}
-
 func TestBuildTimestamp(t *testing.T) {
 	t.Parallel()
